@@ -24,24 +24,30 @@
     "build-prod": "webpack --mode production --env prod",
     "dev": "cd .. && bash ./build-and-install.sh -lf -nb",
     "launch": "webpack-dev-server --open --mode development --env dev --ssl true",
-    "lint": "tslint --project tsconfig.json"
+    "lint": "npm run lint-main && npm run lint-sw",
+    "lint-main": "tslint --project ./src/main",
+    "lint-sw": "tslint --project ./src/sw"
   },
-  "dependencies": {},
+  "dependencies": {
+    "firebase": "^7.0.0"
+  },
   "devDependencies": {
     "@app/sample-app-shared": "~1.1.0",
-    "@nu-art/db-api-generator": "~0.21.0",
-    "@nu-art/firebase": "~0.21.0",
-    "@nu-art/live-docs": "~0.21.0",
-    "@nu-art/permissions": "~0.21.0",
-    "@nu-art/thunderstorm": "~0.21.0",
-    "@nu-art/ts-common": "~0.21.0",
-    "@nu-art/user-account": "~0.21.0",
+    "@nu-art/bug-report": "~0.23.0",
+    "@nu-art/db-api-generator": "~0.23.0",
+    "@nu-art/firebase": "~0.23.0",
+    "@nu-art/live-docs": "~0.23.0",
+    "@nu-art/permissions": "~0.23.0",
+    "@nu-art/push-pub-sub": "~0.23.0",
+    "@nu-art/thunderstorm": "~0.23.0",
+    "@nu-art/ts-common": "~0.23.0",
+    "@nu-art/user-account": "~0.23.0",
     "@types/history": "^4.7.2",
     "@types/qs": "^6.5.2",
     "@types/react": "^16.8.7",
     "@types/react-dom": "^16.8.2",
-    "@types/react-router": "^4.4.5",
-    "@types/react-router-dom": "^4.3.1",
+    "@types/react-router": "^5.1.5",
+    "@types/react-router-dom": "^5.1.3",
     "@types/react-select": "^3.0.8",
     "autoprefixer": "^9.1.5",
     "awesome-typescript-loader": "^5.2.1",
@@ -68,14 +74,14 @@
     "react-color": "^2.14.1",
     "react-dom": "^16.8.1",
     "react-hot-loader": "^4.3.11",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
     "react-select": "^3.1.0",
     "sass-loader": "^7.1.0",
     "source-map-loader": "^0.2.4",
     "style-loader": "^0.23.0",
     "ts-loader": "^5.3.3",
-    "typescript": "^3.3.3333",
+    "typescript": "^3.8.3",
     "url-loader": "^1.1.1",
     "webpack": "^4.20.2",
     "webpack-cli": "^3.1.1",
