@@ -21,9 +21,10 @@ import {
 	ServerApi_Get,
 } from "@nu-art/thunderstorm/backend";
 
-import * as express from "express";
+
 import {ExampleModule} from "@modules/ExampleModule";
 import {ExampleApiGetType} from "@app/sample-app-shared";
+import {ExpressRequest} from "@nu-art/thunderstorm/backend";
 
 class ServerApi_EndpointExample
 	extends ServerApi_Get<ExampleApiGetType> {
@@ -32,7 +33,7 @@ class ServerApi_EndpointExample
 		super("endpoint-example");
 	}
 
-	protected async process(request: express.Request, response: ApiResponse, queryParams: {}, body: void) {
+	protected async process(request: ExpressRequest, response: ApiResponse, queryParams: {}, body: void) {
 		return ExampleModule.getRandomString();
 	}
 }
