@@ -20,8 +20,9 @@ import {
 	ServerApi_Get,
 } from "@nu-art/thunderstorm/backend";
 
-import * as express from "express";
+
 import {ApiType_GetWithoutParams} from "@app/sample-app-shared";
+import {ExpressRequest} from "@nu-art/thunderstorm/backend";
 
 class ServerApi_EndpointExample
 	extends ServerApi_Get<ApiType_GetWithoutParams> {
@@ -30,7 +31,7 @@ class ServerApi_EndpointExample
 		super("get-without-params-endpoint");
 	}
 
-	protected async process(request: express.Request, response: ApiResponse, queryParams: {}, body: void) {
+	protected async process(request: ExpressRequest, response: ApiResponse, queryParams: {}, body: void) {
 		return "another endpoint response"
 	}
 }
