@@ -19,12 +19,11 @@
 // tslint:disable-next-line:no-import-side-effect
 import 'module-alias/register'
 import {
+	ForceUpgrade,
 	RouteResolver,
-	Storm,
-	ForceUpgrade
+	Storm
 } from "@nu-art/thunderstorm/backend";
 import {Environment} from "./config";
-import {ValueChangedListener} from "@modules/ValueChangedListener";
 import {
 	DispatchModule,
 	ExampleModule
@@ -32,10 +31,10 @@ import {
 import {Backend_ModulePack_LiveDocs} from "@nu-art/live-docs/backend";
 import {Module} from "@nu-art/ts-common";
 import {Backend_ModulePack_Permissions} from "@nu-art/permissions/backend";
-import {
-	ProjectBackupScheduler,
-	ProjectFirestoreBackup
-} from "@nu-art/firebase/backend-firestore-backup";
+import {ProjectFirestoreBackup} from "@nu-art/firebase/backend-firestore-backup";
+import {Backend_ModulePack_PushPubSub} from "@nu-art/push-pub-sub/backend";
+import {ValueChangedListener} from "@modules/ValueChangedListener";
+import {Backend_ModulePack_BugReport} from "@nu-art/bug-report/app-backend/core/module-pack";
 
 const functions = require('firebase-functions');
 
@@ -48,13 +47,15 @@ const modules: Module[] = [
 	ForceUpgrade,
 	ProjectFirestoreBackup,
 	// SchedulerExample,
-	ProjectBackupScheduler.setSchedule("every 10 min"),
+	// ProjectBackupScheduler.setSchedule("every 10 min"),
 	DispatchModule
 ];
 
 const _exports = new Storm()
 	.addModules(...Backend_ModulePack_LiveDocs)
 	.addModules(...Backend_ModulePack_Permissions)
+	.addModules(...Backend_ModulePack_PushPubSub)
+	.addModules(...Backend_ModulePack_BugReport)
 	.addModules(...modules)
 	.setInitialRouteResolver(new RouteResolver(require, __dirname, "api"))
 	.setInitialRoutePath("/api")
