@@ -42,10 +42,7 @@ export class Hello
 
 	render() {
 		return <>
-			<div className="ll_h_c"><h1 onClick={() => {
-				ExampleModule.getMessageFromServer1()
-				ExampleModule.getMessageFromServer1()
-			}}>{this.state.label}</h1><LiveDoc docKey="one-mouse-click"/></div>
+			<div className="ll_h_c"><h1 onClick={ExampleModule.getMessageFromServer1}>{this.state.label}</h1><LiveDoc docKey="one-mouse-click"/></div>
 			<div className="ll_h_c"><h1 onDoubleClick={ExampleModule.getMessageFromServer2}>Double click me</h1><LiveDoc docKey="double-mouse-click"/></div>
 			<div className="ll_h_c"><h1 onClick={ForceUpgrade.compareVersion}>Assert version</h1><LiveDoc docKey="assert-version"/></div>
 		</>;
