@@ -17,16 +17,22 @@
  */
 
 import * as React from 'react';
+import {
+	Playground,
+	PlaygroundScreen
+} from "@nu-art/thunderstorm/app-frontend/components/Playground";
 import {Page_DialogExamples} from "./Page_DialogExamples";
 import {Page_ToasterExample} from "./Page_ToasterExample";
 import {Page_ApiGen} from "./Page_ApiGen";
 import {Hello} from "../Hello";
 import {CustomErrorExample} from './CustomErrorExample';
+import {TabsPlayground} from "./TabsPlayground";
+import {GenericSelectPlayground} from "./GenericSelectPlayground";
+import {GenericTabsPlayground} from "./GenericTabsPlayground";
+import { Page_DropDownExamples } from './Page_DropDownExamples';
+import { Page_SpecialKeyboardListener } from './Page_SpecialKeyboardListener';
 import {unitStyle} from "../ui/SelectStyle";
-import {
-	Playground,
-	PlaygroundScreen
-} from "@nu-art/thunderstorm/app-frontend/components/Playground";
+import {MenuPlayground} from "./MenuPlayground";
 
 const icon__arrowClose = require('@res/images/icon__arrowClose.svg');
 const icon__arrowOpen = require('@res/images/icon__arrowOpen.svg');
@@ -81,6 +87,42 @@ export class SamplePlayground extends React.Component<{}> {
 				name: "Custom error",
 				getNode: () => {
 					return <CustomErrorExample/>;
+				}
+			},
+			{
+				name: "Tabs",
+				getNode: () => {
+					return <TabsPlayground/>;
+				}
+			},
+			{
+				name: "GenericTabs",
+				getNode: () => {
+					return <GenericTabsPlayground/>;
+				}
+			},
+			{
+				name: "GenericSelect",
+				getNode: () => {
+					return <GenericSelectPlayground/>;
+				}
+			},
+			{
+				name: "DropDown Examples",
+				getNode: () => {
+					return <Page_DropDownExamples/>;
+				}
+			},
+			{
+				name: "Menu",
+				getNode: () => {
+					return <MenuPlayground/>;
+				}
+			},
+			{
+				name: "Special keyboard listener",
+				getNode: () => {
+					return <Page_SpecialKeyboardListener/>;
 				}
 			},
 		];
