@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-// tslint:disable-next-line:no-import-side-effect
+// tslint:disable:no-import-side-effect
 import './res/styles/styles.scss';
 
 import * as React from 'react';
@@ -32,11 +32,19 @@ import {
 	showEditModalExample
 } from "@nu-art/live-docs/frontend";
 import {ExampleModule} from "@modules/ExampleModule";
+import {PushPubSubModule} from "@nu-art/push-pub-sub/frontend";
+import {BugReportModule} from "@nu-art/bug-report/frontend";
+
+const modules = [
+	ForceUpgrade,
+	PushPubSubModule,
+	ExampleModule
+];
 
 new Thunder()
 	.setConfig(require("./config").config)
 	.addModules(...Frontend_ModulePack_LiveDocs)
-	.addModules(ForceUpgrade, ExampleModule)
+	.addModules(...modules,BugReportModule)
 	.setMainApp(App)
 	.build();
 
